Updated PhysicsCollide callback.
When a vehicle collides with a player the players velocity should be reset, a variable will be applied on the player telling the engine not to apply damage for that tick.

Suspected issues: I am worried that cars may just go through players even when Vehicle Collision is disabled. Not sure if the call order was changed for the collision callback either so damage may still apply. But this is why we're testing it.